# ️ 🙉  improve GetMime performance