… fastmap.go