NEW:
```go
Benchmark_GetMIME/fiber-12         	28303467	        40.9 ns/op	       0 B/op	       0 allocs/op
Benchmark_GetMIME/fiber-12         	27966982	        41.1 ns/op	       0 B/op	       0 allocs/op
Benchmark_GetMIME/fiber-12         	28572391	        41.0 ns/op	       0 B/op	       0 allocs/op
```

OLD:
```go
Benchmark_GetMIME/fiber-12         	17397422	        69.1 ns/op	       0 B/op	       0 allocs/op
Benchmark_GetMIME/fiber-12         	17423786	        69.1 ns/op	       0 B/op	       0 allocs/op
Benchmark_GetMIME/fiber-12         	17338098	        68.8 ns/op	       0 B/op	       0 allocs/op
```