@ReneWerner87 2020-07-22
The algorithm turns a string into a number that is smaller than 10000, i.e. but there can be keys that have the same integer value as e.g. "html", should be rare...
  - >Solution is not safe